Fixes failure to launch "new user" / "popular" screens post onboard
This is a fix for the following issue:

https://meta.getaether.net/t/aether-fails-to-show-new-user-modal-after-saying-yes-to-onboarding-question-create-new-user/350

The problem was that a navigation guard was in place preventing navigation outside onboarding screens until onboarding was set to complete in VueX store.

This fix ensures that router.push is called only *after* onboarding is successfully set to complete.

There is a slight issue with scroll position in the "create new user" screen, where the screen's title is not visible, but it looks alright and feels quite natural at the default window size so I think it's fine.

It might be good to make a mixin for resetting the scroll position and use that here just to be sure.